refactor: :recycle: tucks away bus logic in custom hook
